6/7 done.

+++ app/Http/Controllers/AdminProductController.php
@@ -108,6 +108,8 @@ class AdminProductController extends Controller
      */
     public function edit(Product $product)
     {
+    	if($product->view_time <= Carbon::now())
+    		return redirect()->back()->with('has_begin', '已過公布時間，無法編輯');
 		$files = get_files(storage_path('app/public/products/'.$product->id));
 		$pics[$product->id] = $files;
 		
@@ -128,6 +130,8 @@ class AdminProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+    	if($product->view_time >= Carbon::now())
+    		return redirect()->back()->with('has_begin', '已過公布時間，無法編輯');
 		$att['name'] = $request->input('name');
 		$att['view_time'] = $request->input('view_time');
 		$att['start_time'] = $request->input('start_time');
@@ -166,6 +170,8 @@ class AdminProductController extends Controller
 
     public function destroy(Product $product)
     {
+    	if($product->view_time <= Carbon::now())
+    		return redirect()->back()->with('has_begin', '已過公布時間，無法刪除');
 		Storage::deleteDirectory("public/products/".$product->id);
         $product->delete(); 
         return redirect()->route('adm_Product');
